+++ qualtran/bloqs/multiplexers/__init__.py
@@ -1,0 +1,22 @@
+#  Copyright 2023 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+r"""Multiplexed Bloqs that can be used to apply a coherent "for-loop" via unary iteration.
+
+Unary iteration can be used to implement unitaries of the form
+
+ $$
+     U = \sum_{l=0}^{2^n-1} U_{l}|l\rangle \langle l|
+ $$
+"""
