+++ qualtran/bloqs/multiplexers/apply_gate_to_lth_target_test.py
@@ -18,7 +18,10 @@ import pytest
 import qualtran.testing as qlt_testing
 from qualtran import BoundedQUInt, Register, Signature
 from qualtran._infra.gate_with_registers import get_named_qubits, total_bits
-from qualtran.bloqs.apply_gate_to_lth_target import _apply_z_to_odd, ApplyGateToLthQubit
+from qualtran.bloqs.multiplexers.apply_gate_to_lth_target import (
+    _apply_z_to_odd,
+    ApplyGateToLthQubit,
+)
 from qualtran.cirq_interop.bit_tools import iter_bits
 from qualtran.cirq_interop.testing import assert_circuit_inp_out_cirqsim, GateHelper
