+++ qualtran/bloqs/multiplexers/select_pauli_lcu_test.py
@@ -18,7 +18,7 @@ import numpy as np
 import pytest
 
 from qualtran._infra.gate_with_registers import get_named_qubits
-from qualtran.bloqs.select_pauli_lcu import SelectPauliLCU
+from qualtran.bloqs.multiplexers.select_pauli_lcu import SelectPauliLCU
 from qualtran.cirq_interop.bit_tools import iter_bits
 from qualtran.cirq_interop.testing import assert_circuit_inp_out_cirqsim
 from qualtran.testing import assert_valid_bloq_decomposition, execute_notebook
