+++ qualtran/bloqs/qubitization_walk_operator_test.py
@@ -20,8 +20,8 @@ from qualtran._infra.gate_with_registers import get_named_qubits, total_bits
 from qualtran.bloqs.mcmt.multi_control_multi_target_pauli import MultiControlPauli
 from qualtran.bloqs.qubitization_walk_operator import QubitizationWalkOperator
 from qualtran.bloqs.reflection_using_prepare_test import construct_gate_helper_and_qubit_order
-from qualtran.bloqs.select_pauli_lcu import SelectPauliLCU
-from qualtran.bloqs.select_pauli_lcu_test import get_1d_Ising_hamiltonian
+from qualtran.bloqs.multiplexers.select_pauli_lcu import SelectPauliLCU
+from qualtran.bloqs.multiplexers.select_pauli_lcu_test import get_1d_Ising_hamiltonian
 from qualtran.bloqs.state_preparation import StatePreparationAliasSampling
 from qualtran.testing import assert_valid_bloq_decomposition, execute_notebook
